refactor: decouple logic from app to children
