+++ src/app.tsx
@@ -1,14 +1,17 @@
-import debounce from 'lodash/debounce'
-import times from 'lodash/times'
 import { Fabric, initializeIcons, loadTheme } from 'office-ui-fabric-react'
-import { rgba } from 'polished'
-import random from 'random'
-import React, { useCallback, useEffect, useRef, useState } from 'react'
+import React, { useEffect, useReducer } from 'react'
 import { useTranslation } from 'react-i18next'
 import styled, { createGlobalStyle, ThemeProvider } from 'styled-components'
 
-import { darkTheme, getLocaleFontFamily, lightTheme } from './theme'
+import {
+  darkTheme,
+  DispatchThemeChangeContext,
+  getLocaleFontFamily,
+  lightTheme,
+  ThemeIsDarkContext,
+} from './theme'
 
+import { Background } from './components/background'
 import { Content } from './components/content'
 import { Footer } from './components/footer'
 import { Header } from './components/header'
@@ -31,15 +34,6 @@ const Container = styled(Fabric)`
   flex-direction: column;
 `
 
-const Canvas = styled.canvas`
-  z-index: -1;
-  position: fixed;
-  top: 0;
-  left: 0;
-  background-color: ${props => props.theme.palette.white};
-  transition: 0.3s;
-`
-
 const Wrapper = styled.div`
   color: ${props => props.theme.palette.neutralPrimary};
   flex: 1;
@@ -48,55 +42,6 @@ const Wrapper = styled.div`
   flex-direction: column;
 `
 
-/**
- * Draws an hexagone
- * @param ctx canvas context
- * @param x x coordinate
- * @param y y coordinate
- * @param r radius or edge length
- * @param fillStyle fill style
- * @param strokeStyle stroke style
- */
-const drawHexagone = (
-  ctx: CanvasRenderingContext2D,
-  x: number,
-  y: number,
-  r: number,
-  fillStyle: CanvasFillStrokeStyles['fillStyle'],
-  strokeStyle: CanvasFillStrokeStyles['strokeStyle'],
-) => {
-  ctx.beginPath()
-
-  ctx.moveTo(x, y - r)
-
-  const unit = Math.PI / 3
-  // anti-clockwise
-  times(6, (n: number) => {
-    const angle = Math.PI / 2 + unit * (n + 1)
-    ctx.lineTo(x + r * Math.cos(angle), y - r * Math.sin(angle))
-  })
-  ctx.closePath()
-  ctx.fillStyle = fillStyle
-  ctx.strokeStyle = strokeStyle
-  ctx.lineWidth = 1
-  ctx.fill()
-  ctx.stroke()
-}
-
-const normal = random.normal(0.75, 0.25)
-
-/**
- * get a random x,y coordinate with provide width and height
- * @param w width
- * @param h height
- */
-const getRandomXY = (w: number, h: number) => {
-  const x = Math.floor(w * Math.random())
-  const y = Math.floor(h * normal())
-
-  return { x, y }
-}
-
 const choiceDark = localStorage.getItem('theme') === 'dark'
 
 export const App = () => {
@@ -104,66 +49,10 @@ export const App = () => {
 
   const fontFamily = getLocaleFontFamily(i18n.language)
 
-  useEffect(() => {
-    loadTheme({
-      ...(isDark ? darkTheme : lightTheme),
-      defaultFontStyle: {
-        fontFamily,
-      },
-    })
-  }, [i18n.language])
-
-  const [isDark, setIsDark] = useState(choiceDark)
-
-  const canvas = useRef<HTMLCanvasElement>(null)
-
-  const drawCanvas = useCallback(() => {
-    if (!canvas.current) {
-      return
-    }
-    const cvs = canvas.current
-    const ctx = cvs.getContext('2d')!
-    const pr = window.devicePixelRatio || 1
-    const w = window.innerWidth
-    const h = window.innerHeight
-
-    cvs.width = w * pr
-    cvs.height = h * pr
-    cvs.style.width = `${w}px`
-    cvs.style.height = `${h}px`
-    ctx.scale(pr, pr)
-    ctx.clearRect(0, 0, w, h)
-
-    times(30, () => {
-      const { x, y } = getRandomXY(w, h)
-
-      drawHexagone(
-        ctx,
-        x,
-        y,
-        50,
-        `${rgba((isDark ? darkTheme : lightTheme).palette.black, 0.1)}`,
-        'transparent',
-      )
-    })
-    times(30, () => {
-      const { x, y } = getRandomXY(w, h)
-
-      drawHexagone(
-        ctx,
-        x,
-        y,
-        50,
-        'transparent',
-        `${rgba((isDark ? darkTheme : lightTheme).palette.black, 0.1)}`,
-      )
-    })
-  }, [canvas, isDark])
-
-  useEffect(() => {
-    drawCanvas()
-    window.addEventListener('resize', debounce(drawCanvas, 100))
-  }, [])
+  const [isDark, dispatch] = useReducer(
+    (state: boolean, value: any) => !state,
+    choiceDark,
+  )
 
   useEffect(() => {
     loadTheme({
@@ -172,21 +61,23 @@ export const App = () => {
         fontFamily,
       },
     })
-    drawCanvas()
-    localStorage.setItem('theme', isDark ? 'dark' : 'light')
-  }, [isDark])
+  }, [i18n.language, isDark])
 
   return (
     <ThemeProvider theme={isDark ? darkTheme : lightTheme}>
-      <Container>
-        <GlobalStyle />
-        <Canvas ref={canvas} />
-        <Wrapper>
-          <Header onChangeTheme={() => setIsDark(!isDark)} isDark={isDark} />
-          <Content />
-          <Footer />
-        </Wrapper>
-      </Container>
+      <DispatchThemeChangeContext.Provider value={dispatch}>
+        <ThemeIsDarkContext.Provider value={isDark}>
+          <GlobalStyle />
+          <Background />
+          <Container>
+            <Wrapper>
+              <Header />
+              <Content />
+              <Footer />
+            </Wrapper>
+          </Container>
+        </ThemeIsDarkContext.Provider>
+      </DispatchThemeChangeContext.Provider>
     </ThemeProvider>
   )
 }
